core: quantize octahedral normals in double, as upstream does
The rounding rule already matched upstream. Only the width of the
intermediate differed: C++ projects onto the octahedron in double and
floors a double, while this computed the whole thing in f32. That decides
the result whenever the floored value lands exactly on .5, which ordinary
normals do all the time -- (0, 0.7071, 0.7071) gives exactly 511.5 at ten
bits -- and the two encoders then chose neighbouring octahedral
coordinates for identical input.

The symptom was mirror-image output, which is what gave it away:

  input   [0.0000, 0.7071, 0.7071]
  ours    [0.0000, 0.7057, 0.7085]
  theirs  [0.0000, 0.7085, 0.7057]

Byte parity with the C++ encoder on attributed meshes goes from 24 of 55
cases to 38, and every remaining difference is at speeds 0 to 3, where two
prediction schemes nothing above speed 3 uses are still to be reconciled.
The new test asserts from speed 4 up and reports the rest, so the range
that is right stays guarded while that is open. A second test compares the
two decoders on one payload, pairing points by position because they
number points differently once edgebreaker is in play; it asserts from
speed 1, since at speed 0 the same point still comes back with a different
normal.

+++ crates/draco-core/src/normal_compression_utils.rs
@@ -204,15 +204,21 @@ impl OctahedronToolBox {
     }
 
     pub fn float_vector_to_quantized_octahedral_coords(&self, vector: &[f32; 3]) -> (i32, i32) {
-        let abs_sum = vector[0].abs() + vector[1].abs() + vector[2].abs();
+        // Double, as upstream does it, and not for extra accuracy: the value
+        // being floored lands exactly on `.5` for ordinary normals -- (0, k, k)
+        // among them -- and single precision resolves that tie the other way.
+        // The two encoders then pick neighbouring octahedral coordinates for
+        // the same input.
+        let abs_sum =
+            (vector[0] as f64).abs() + (vector[1] as f64).abs() + (vector[2] as f64).abs();
 
         // Adjust values such that abs sum equals 1.
-        let mut scaled_vector = [0.0; 3];
+        let mut scaled_vector = [0.0f64; 3];
         if abs_sum > 1e-6 {
             let scale = 1.0 / abs_sum;
-            scaled_vector[0] = vector[0] * scale;
-            scaled_vector[1] = vector[1] * scale;
-            scaled_vector[2] = vector[2] * scale;
+            scaled_vector[0] = vector[0] as f64 * scale;
+            scaled_vector[1] = vector[1] as f64 * scale;
+            scaled_vector[2] = vector[2] as f64 * scale;
         } else {
             scaled_vector[0] = 1.0;
             scaled_vector[1] = 0.0;
@@ -221,8 +227,8 @@ impl OctahedronToolBox {
 
         // Scale vector such that the sum equals the center value.
         let mut int_vec = [0; 3];
-        int_vec[0] = (scaled_vector[0] * self.center_value as f32 + 0.5).floor() as i32;
-        int_vec[1] = (scaled_vector[1] * self.center_value as f32 + 0.5).floor() as i32;
+        int_vec[0] = (scaled_vector[0] * self.center_value as f64 + 0.5).floor() as i32;
+        int_vec[1] = (scaled_vector[1] * self.center_value as f64 + 0.5).floor() as i32;
 
         // Make sure the sum is exactly the center value.
         int_vec[2] = self.center_value - int_vec[0].abs() - int_vec[1].abs();
